+++ CHANGELOG.md
@@ -9,6 +9,7 @@ with the current date and the next changes should go under a **[Next]** header.
 
 ## 28 March 2018
 
+* Show queue name and location on queue page ([@sgorse](https://github.com/sgoese) in [#81](https://github.com/illinois/queue/pull/81) )
 * Improve spacing of course buttons on homepage. ([@nwalters512](https://github.com/nwalters512) in [#60](https://github.com/illinois/queue/pull/60))
 * Add support and tooling for Sequelize database migrations. ([@nwalters512](https://github.com/nwalters512) in [#56](https://github.com/illinois/queue/pull/56))
 * Add confirmation prompt for cancelling and deleting student questions. ([@muakasan](https://github.com/muakasan) in [#71](https://github.com/illinois/queue/pull/71))